Добавлено класс для работы с сервером. Реализованы посылки типа 1 (запрос на подключение).

+++ etcp.h
@@ -29,8 +29,8 @@ extern "C" {
 typedef void ( *tofunc_t )(void *);
 
 void error(int, int, const char *, ...);
-int readn(SOCKET, char *, size_t);
-int readvrec(SOCKET, char *, size_t);
+int readn(SOCKET, char *, int);
+int readvrec(SOCKET, char *, int);
 //int readcrlf(SOCKET, char*, size_t);
 //int readline(SOCKET, char*, size_t);
 SOCKET tcp_server(char *, char *);
